# go_backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Smart Waste Management API
  description: |
    RESTful API for an AI & AR-Driven Smart Waste Management Platform.
    
    ## Features
    - **IoT Integration**: Real-time data from ESP32-based smart bins via MQTT
    - **User Management**: Profile and reward point tracking
    - **Driver Management**: Route optimization and task verification
    - **Dashboard**: Company management, pricing rules, and analytics
    - **Valuation Engine**: AI-based waste pricing
    
    ## Authentication
    Authentication endpoints are placeholders. Implement JWT or OAuth2 for production.
  version: 1.0.0
  contact:
    name: Smart Waste Team
    email: support@smartwaste.io
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
...
paths:
...
  /companies:
    get:
      tags:
        - Companies
      summary: List all companies
      responses:
        '200':
          description: List of companies
    post:
      tags:
        - Companies
      summary: Create a new company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompanyRequest'
      responses:
        '201':
          description: Company created
  /companies/{id}:
    get:
      tags:
        - Companies
      summary: Get company by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Company details
    put:
      tags:
        - Companies
      summary: Update company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCompanyRequest'
      responses:
        '200':
          description: Company updated
    delete:
      tags:
        - Companies
      summary: Delete company
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Company deleted

  # Pricing Rules
  /pricing-rules:
    get:
      tags:
        - Pricing Rules
      summary: List all pricing rules
      responses:
        '200':
          description: List of pricing rules
    post:
      tags:
        - Pricing Rules
      summary: Create a new pricing rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePricingRuleRequest'
      responses:
        '201':
          description: Pricing rule created
  /pricing-rules/{id}:
    get:
      tags:
        - Pricing Rules
      summary: Get pricing rule by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Pricing rule details
    put:
      tags:
        - Pricing Rules
      summary: Update pricing rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePricingRuleRequest'
      responses:
        '200':
          description: Pricing rule updated
    delete:
      tags:
        - Pricing Rules
      summary: Delete pricing rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Pricing rule deleted
  /valuations:
    post:
      tags:
        - Pricing Rules
      summary: Calculate waste valuation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValuationRequest'
      responses:
        '200':
          description: Valuation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValuationResponse'

  # Analytics
...
components:
  schemas:
...
    CreateCompanyRequest:
      type: object
      required:
        - name
        - email
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string
        registration_number:
          type: string
    UpdateCompanyRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        country:
          type: string
        is_active:
          type: boolean
    CreatePricingRuleRequest:
      type: object
      required:
        - waste_type
        - condition
        - price_per_kg
        - currency
      properties:
        waste_type:
          type: string
        condition:
          type: string
        price_per_kg:
          type: number
        currency:
          type: string
          maxLength: 3
        min_weight_kg:
          type: number
        max_weight_kg:
          type: number
        company_id:
          type: string
          format: uuid
    UpdatePricingRuleRequest:
      type: object
      properties:
        waste_type:
          type: string
        condition:
          type: string
        price_per_kg:
          type: number
        currency:
          type: string
        is_active:
          type: boolean
    ValuationRequest:
      type: object
      required:
        - waste_type
        - condition
        - weight_kg
      properties:
        waste_type:
          type: string
        condition:
          type: string
        weight_kg:
          type: number
    ValuationResponse:
      type: object
      properties:
        waste_type:
          type: string
        condition:
          type: string
        weight_kg:
          type: number
        price_per_kg:
          type: number
        total_price:
          type: number
        currency:
          type: string
        message:
          type: string
...